+++ rescue/src/lib.rs
@@ -1,8 +1,8 @@
 #![allow(dead_code)] // TODO: remove when we settle on implementation details and publicly export
 
-mod inverse_sbox;
 mod rescue;
+mod sbox;
 mod util;
 
-pub use inverse_sbox::*;
 pub use rescue::*;
+pub use sbox::*;
